Move new check code to old check.py file
