+++ pyvista/core/input_validation/check.py
@@ -1,36 +1,60 @@
 """Functions that check the type and/or value of inputs.
 
-.. versionadded:: 0.43.0
+.. versionadded:: 0.44.0
 
-A ``check`` function typically:
+An array checker function typically:
 
-* Performs a simple validation on a single input variable.
+* Checks the type and/or value of a single input variable.
 * Raises an error if the check fails due to invalid input.
 * Does not modify input or return anything.
 
 """
-from collections.abc import Iterable, Sequence
-from numbers import Number
-from typing import Tuple, Union, get_args, get_origin
+from numbers import Number, Real
+from typing import (
+    Any,
+    Iterable,
+    List,
+    Literal,
+    Optional,
+    Sequence,
+    Sized,
+    Tuple,
+    TypeVar,
+    Union,
+    cast,
+    get_args,
+    get_origin,
+)
 
 import numpy as np
 
+from pyvista.core._typing_core import Array, NumpyArray, Vector
+from pyvista.core._typing_core._array_like import _ArrayLikeOrScalar, _NumberType
 from pyvista.core.utilities.arrays import cast_to_ndarray
+from pyvista.core.validation._array_like import DTypeLike, Shape, ShapeLike, _array_like_props
+
+_T = TypeVar('_T')
 
 
-def check_is_subdtype(arg1, arg2, /, *, name='Input'):
-    """Check if a data-type is a subtype of another data-type(s).
+def check_subdtype(
+    input_obj: Union[DTypeLike, _ArrayLikeOrScalar[_NumberType]],
+    base_dtype: Union[DTypeLike, Tuple[DTypeLike, ...], List[DTypeLike]],
+    /,
+    *,
+    name: str = 'Input',
+):
+    """Check if an input's data-type is a subtype of another data-type(s).
 
     Parameters
     ----------
-    arg1 : dtype_like | array_like
-        ``dtype`` or object coercible to one. If ``array_like``, the dtype
-        of the array is used.
+    input_obj : DType | Array | Scalar
+        ``dtype`` object (or object coercible to one) or an array-like object.
+        If array-like, the dtype of the array is used.
 
-    arg2 : dtype_like | list[dtype_like]
-        ``dtype``-like object or a list of ``dtype``-like objects.
-        If a list, ``arg1`` must be a subtype of at least one of the
-        specified dtypes.
+    base_dtype : DType | Sequence[DType]
+        ``dtype``-like object or a sequence of ``dtype``-like objects. The ``input_obj``
+        must be a subtype of this value. If a sequence, ``input_obj`` must be a
+        subtype of at least one of the specified dtypes.
 
     name : str, default: "Input"
         Variable name to use in the error messages if any are raised.
@@ -38,140 +62,112 @@ def check_is_subdtype(arg1, arg2, /, *, name='Input'):
     Raises
     ------
     TypeError
-        If ``arg1`` is not a subtype of ``arg2``.
+        If ``input_obj`` is not a subtype of ``base_dtype``.
 
     See Also
     --------
-    check_is_real
-    check_is_number
+    check_real
+    check_number
 
     Examples
     --------
-    Check if ``int`` is a subtype of ``np.integer``.
+    Check if ``float`` is a subtype of ``np.floating``.
 
     >>> import numpy as np
-    >>> import pyvista.core.input_validation as valid
-    >>> valid.check_is_subdtype(float, np.floating)
+    >>> from pyvista.core import validation
+    >>> validation.check_subdtype(float, np.floating)
 
     Check from multiple allowable dtypes.
 
-    >>> valid.check_is_subdtype(int, [np.integer, np.floating])
+    >>> validation.check_subdtype(int, [np.integer, np.floating])
 
     Check an array's dtype.
 
     >>> arr = np.array([1, 2, 3], dtype='uint8')
-    >>> valid.check_is_subdtype(arr, np.integer)
+    >>> validation.check_subdtype(arr, np.integer)
 
     """
-    if isinstance(arg1, np.dtype):
-        pass
-    elif isinstance(arg1, np.ndarray):
-        arg1 = arg1.dtype
+    input_dtype: DTypeLike
+    if isinstance(input_obj, (tuple, list, np.ndarray)):
+        _, _, input_dtype, _ = _array_like_props(input_obj)
     else:
-        arg1 = np.dtype(arg1)
+        input_dtype = cast(DTypeLike, input_obj)
 
-    if not isinstance(arg2, (list, tuple)):
-        arg2 = [arg2]
-    for d in arg2:
-        check_is_dtypelike(d)
-        if np.issubdtype(arg1, d):
-            return
-    msg = f"{name} has incorrect dtype of '{arg1}'. "
-    if len(arg2) == 1:
-        msg += f"The dtype must be a subtype of {arg2[0]}."
+    if not isinstance(base_dtype, (tuple, list)):
+        base_dtype = [base_dtype]
+
+    if any(np.issubdtype(input_dtype, base) for base in base_dtype):
+        return None
     else:
-        msg += f"The dtype must be a subtype of at least one of \n{arg2}."
-    raise TypeError(msg)
+        # Not a subdtype, so raise error
+        msg = f"{name} has incorrect dtype of {input_dtype}. "
+        if len(base_dtype) == 1:
+            msg += f"The dtype must be a subtype of {base_dtype[0]}."
+        else:
+            msg += f"The dtype must be a subtype of at least one of \n{base_dtype}."
+        raise TypeError(msg)
 
 
-def check_is_dtypelike(dtype):
-    """Check if an input is dtype-like.
+def check_numeric(arr: _ArrayLikeOrScalar, /, *, name: str = "Array"):
+    # TODO: rename and modify as 'check_complex'
+    """Check if an array is float, integer, or complex type.
+
+    Notes
+    -----
+    Arrays with ``infinity`` or ``NaN`` values are numeric  and
+    will not raise an error. Use :func:`check_finite` to check for
+    finite values.
 
     Parameters
     ----------
-    dtype : dtype_like
-        ``dtype`` or object coercible to one.
+    arr : float | Array[float]
+        Number or array to check.
+
+    name : str, default: "Array"
+        Variable name to use in the error messages if any are raised.
 
     Raises
     ------
     TypeError
-        If input is not coercible to a dtype object.
+        If the array is not numeric.
 
     See Also
     --------
-    check_is_arraylike
+    check_real
+    check_finite
 
     Examples
     --------
-    Check if an input is dtype-like.
+    Check if an array is numeric.
 
-    >>> import numpy as np
-    >>> import pyvista.core.input_validation as valid
-    >>> valid.check_is_dtypelike(float)
-    >>> valid.check_is_dtypelike(np.dtype(np.integer))
-    >>> valid.check_is_dtypelike('uint8')
+    >>> from pyvista.core import validation
+    >>> validation.check_numeric([1, 2.0, 3 + 3j])
 
     """
-    # Return early for common dtype cases
-    if dtype not in [np.integer, np.floating, np.number, int, float] and not isinstance(
-        dtype, np.dtype
-    ):
-        try:
-            np.dtype(dtype)
-        except TypeError as e:
-            raise TypeError(f"'{dtype}' is not a valid NumPy data type.") from e
+    arr = arr if isinstance(arr, np.ndarray) else cast_to_ndarray(arr)
 
-
-def check_is_arraylike(arr):
-    """Check if an input can be cast as a NumPy ndarray.
-
-    Notes
-    -----
-    This check is done by calling :func:`~pyvista.core.utilities.arrays.cast_to_ndarray`
-    internally. Use that function directly if the cast array is needed
-    for further processing.
-
-    Parameters
-    ----------
-    arr : array_like
-        Array to check.
-
-    Raises
-    ------
-    ValueError
-        If input cannot be cast as a NumPy ndarray.
-
-    See Also
-    --------
-    check_is_dtypelike
-
-    Examples
-    --------
-    Check if an input is array-like.
-
-    >>> import pyvista.core.input_validation as valid
-    >>> valid.check_is_arraylike([1, 2, 3])
-
-    """
+    # Return early for common cases
+    if arr.dtype.type in [np.int32, np.int64, np.float32, np.float64, np.complex_]:
+        return
     try:
-        cast_to_ndarray(arr) if not isinstance(arr, np.ndarray) else None
-    except ValueError:
-        raise
+        check_subdtype(arr, np.number, name=name)
+    except TypeError as e:
+        raise TypeError(f"{name} must be numeric.") from e
 
 
-def check_is_real(arr, /, *, name="Array"):
+def check_real(arr: _ArrayLikeOrScalar[_NumberType], /, *, name: str = "Array"):
     """Check if an array has real numbers, i.e. float or integer type.
 
     Notes
     -----
     Arrays with ``infinity`` or ``NaN`` values are considered real and
-    will not raise an error. Use :func:`check_is_finite` to check for
+    will not raise an error. Use :func:`check_finite` to check for
     finite values.
 
     Parameters
     ----------
-    arr : array_like
-        Array to check.
+    arr : float | Array[float] | NDArray[float]
+        Number or array to check.
 
     name : str, default: "Array"
         Variable name to use in the error messages if any are raised.
@@ -183,14 +179,18 @@ def check_is_real(arr, /, *, name="Array"):
 
     See Also
     --------
-    check_is_finite
+    check_numeric
+        Similar function which allows complex numbers.
+    check_scalar
+        Similar function for a single number or 0-dimensional ndarrays.
+    check_finite
 
     Examples
     --------
     Check if an array has real numbers.
 
-    >>> import pyvista.core.input_validation as valid
-    >>> valid.check_is_real([1, 2, 3])
+    >>> from pyvista.core import validation
+    >>> validation.check_real([1, 2, 3])
 
     """
     arr = arr if isinstance(arr, np.ndarray) else cast_to_ndarray(arr)
@@ -202,18 +202,26 @@ def check_is_real(arr, /, *, name="Array"):
     # Do not use np.isreal as it will fail in some cases (e.g. scalars).
     # Check dtype directly instead
     try:
-        check_is_subdtype(arr, (np.floating, np.integer), name=name)
+        check_subdtype(arr, (np.floating, np.integer), name=name)
     except TypeError as e:
         raise TypeError(f"{name} must have real numbers.") from e
 
 
-def check_is_sorted(arr, /, *, ascending=True, strict=False, axis=-1, name="Array"):
+def check_sorted(
+    arr: Union[_NumberType, Array[_NumberType]],
+    /,
+    *,
+    ascending: bool = True,
+    strict: bool = False,
+    axis: int = -1,
+    name: str = "Array",
+):
     """Check if an array's values are sorted.
 
     Parameters
     ----------
-    arr : array_like
-        Array to check.
+    arr : float | Array[float]
+        Number or array to check.
 
     ascending : bool, default: True
         If ``True``, check if the array's elements are in ascending order.
@@ -226,8 +234,9 @@ def check_is_sorted(arr, /, *, ascending=True, strict=False, axis=-1, name="Arra
         must be unique.
 
     axis : int | None, default: -1
-        Axis along which to sort. If ``None``, the array is flattened before
-        sorting. The default is ``-1``, which sorts along the last axis.
+        Axis along which to check sorting. If ``None``, the array is flattened
+        before checking. The default is ``-1``, which checks sorting along the
+        last axis.
 
     name : str, default: "Array"
         Variable name to use in the error messages if any are raised.
@@ -239,14 +248,14 @@ def check_is_sorted(arr, /, *, ascending=True, strict=False, axis=-1, name="Arra
 
     See Also
     --------
-    check_is_in_range
+    check_range
 
     Examples
     --------
     Check if an array's values are sorted,
 
-    >>> import pyvista.core.input_validation as valid
-    >>> valid.check_is_sorted([1, 2, 3])
+    >>> from pyvista.core import validation
+    >>> validation.check_sorted([1, 2, 3])
 
     """
     arr = arr if isinstance(arr, np.ndarray) else cast_to_ndarray(arr)
@@ -263,11 +272,11 @@ def check_is_sorted(arr, /, *, ascending=True, strict=False, axis=-1, name="Arra
     else:
         if not axis == -1:
             # Validate axis
-            check_is_number(axis, name="Axis")
-            check_is_integerlike(axis, name="Axis")
+            check_number(axis, name="Axis")
+            check_integerlike(axis, name="Axis")
             axis = int(axis)
             try:
-                check_is_in_range(axis, rng=[-arr.ndim, arr.ndim - 1], name="Axis")
+                check_range(axis, rng=[-arr.ndim, arr.ndim - 1], name="Axis")
             except ValueError:
                 raise ValueError(f"Axis {axis} is out of bounds for ndim {arr.ndim}.")
         if axis < 0:
@@ -278,20 +287,20 @@ def check_is_sorted(arr, /, *, ascending=True, strict=False, axis=-1, name="Arra
     # Create two slicers to compare consecutive elements with each other
     first = [slice(None)] * arr.ndim
     first[axis] = slice(None, -1)
-    first = tuple(first)
+    first = tuple(first)  # type: ignore
 
     second = [slice(None)] * arr.ndim
     second[axis] = slice(1, None)
-    second = tuple(second)
+    second = tuple(second)  # type: ignore
 
     if ascending and not strict:
-        is_sorted = np.all(arr[first] <= arr[second])
+        is_sorted = np.all(arr[first] <= arr[second])  # type: ignore
     elif ascending and strict:
-        is_sorted = np.all(arr[first] < arr[second])
+        is_sorted = np.all(arr[first] < arr[second])  # type: ignore
     elif not ascending and not strict:
-        is_sorted = np.all(arr[first] >= arr[second])
+        is_sorted = np.all(arr[first] >= arr[second])  # type: ignore
     else:  # not ascending and strict
-        is_sorted = np.all(arr[first] > arr[second])
+        is_sorted = np.all(arr[first] > arr[second])  # type: ignore
     if not is_sorted:
         if arr.size <= 4:
             # Show the array's elements in error msg if array is small
@@ -299,17 +308,17 @@ def check_is_sorted(arr, /, *, ascending=True, strict=False, axis=-1, name="Arra
         else:
             msg_body = f"with {arr.size} elements"
         order = "ascending" if ascending else "descending"
-        strict = "strict " if strict else ""
+        strict = "strict " if strict else ""  # type: ignore
         raise ValueError(f"{name} {msg_body} must be sorted in {strict}{order} order.")
 
 
-def check_is_finite(arr, /, *, name="Array"):
+def check_finite(arr: Union[_NumberType, Array[_NumberType]], /, *, name: str = "Array"):
     """Check if an array has finite values, i.e. no NaN or Inf values.
 
     Parameters
     ----------
-    arr : array_like
-        Array to check.
+    arr : float | int | bool | Array[float] | Array[int] | Array[bool]
+        Number or array to check.
 
     name : str, default: "Array"
         Variable name to use in the error messages if any are raised.
@@ -321,14 +330,14 @@ def check_is_finite(arr, /, *, name="Array"):
 
     See Also
     --------
-    check_is_real
+    check_real
 
     Examples
     --------
     Check if an array's values are finite.
 
-    >>> import pyvista.core.input_validation as valid
-    >>> valid.check_is_finite([1, 2, 3])
+    >>> from pyvista.core import validation
+    >>> validation.check_finite([1, 2, 3])
 
     """
     arr = arr if isinstance(arr, np.ndarray) else cast_to_ndarray(arr)
@@ -336,13 +345,15 @@ def check_is_finite(arr, /, *, name="Array"):
         raise ValueError(f"{name} must have finite values.")
 
 
-def check_is_integerlike(arr, /, *, strict=False, name="Array"):
+def check_integerlike(
+    arr: Union[_NumberType, Array[_NumberType]], /, *, strict: bool = False, name: str = "Array"
+):
     """Check if an array has integer or integer-like float values.
 
     Parameters
     ----------
-    arr : array_like
-        Array to check.
+    arr : float | Array[float]
+        Number or array to check.
 
     strict : bool, default: False
         If ``True``, the array's data must be a subtype of ``np.integer``
@@ -357,38 +368,38 @@ def check_is_integerlike(arr, /, *, strict=False, name="Array"):
         If any element's value differs from its floor.
 
     TypeError
-        If ``strict=True`` and the array's dtype is not integral.
+        If ``strict=True`` and the array's dtype is not integr
 
     See Also
     --------
-    check_is_nonnegative
-    check_is_subdtype
+    check_nonnegative
+    check_subdtype
 
     Examples
     --------
     Check if an array has integer-like values.
 
-    >>> import pyvista.core.input_validation as valid
-    >>> valid.check_is_integerlike([1.0, 2.0])
+    >>> from pyvista.core import validation
+    >>> validation.check_integerlike([1.0, 2.0])
 
     """
     arr = arr if isinstance(arr, np.ndarray) else cast_to_ndarray(arr)
     if strict:
         try:
-            check_is_subdtype(arr, np.integer)
+            check_subdtype(arr, np.integer)
         except TypeError:
             raise
     elif not np.array_equal(arr, np.floor(arr)):
         raise ValueError(f"{name} must have integer-like values.")
 
 
-def check_is_nonnegative(arr, /, *, name="Array"):
+def check_nonnegative(arr: Union[_NumberType, Array[_NumberType]], /, *, name: str = "Array"):
     """Check if an array's elements are all nonnegative.
 
     Parameters
     ----------
-    arr : array_like
-        Array to check.
+    arr : float | Array[float]
+        Number or array to check.
 
     name : str, default: "Array"
         Variable name to use in the error messages if any are raised.
@@ -400,32 +411,39 @@ def check_is_nonnegative(arr, /, *, name="Array"):
 
     See Also
     --------
-    check_is_greater_than
-    check_is_less_than
+    check_greater_than
+    check_less_than
 
     Examples
     --------
     Check if an array's values are non-negative.
 
-    >>> import pyvista.core.input_validation as valid
-    >>> valid.check_is_nonnegative([1, 2, 3])
+    >>> from pyvista.core import validation
+    >>> validation.check_nonnegative([1, 2, 3])
 
     """
     try:
-        check_is_greater_than(arr, 0, strict=False, name=name)
+        check_greater_than(arr, 0, strict=False, name=name)
     except ValueError:
         raise
 
 
-def check_is_greater_than(arr, /, value, *, strict=True, name="Array"):
+def check_greater_than(
+    arr: Union[_NumberType, Array[_NumberType]],
+    /,
+    value: float,
+    *,
+    strict: bool = True,
+    name: str = "Array",
+):
     """Check if an array's elements are all greater than some value.
 
     Parameters
     ----------
-    arr : array_like
-        Array to check.
+    arr : float | Array[float]
+        Number or array to check.
 
-    value : Number
+    value : float
         Value which the array's elements must be greater than.
 
     strict : bool, default: True
@@ -443,38 +461,43 @@ def check_is_greater_than(arr, /, value, *, strict=True, name="Array"):
 
     See Also
     --------
-    check_is_less_than
-    check_is_in_range
-    check_is_nonnegative
+    check_less_than
+    check_range
+    check_nonnegative
 
     Examples
     --------
     Check if an array's values are greater than 0.
 
-    >>> import pyvista.core.input_validation as valid
-    >>> valid.check_is_greater_than([1, 2, 3], value=0)
+    >>> from pyvista.core import validation
+    >>> validation.check_greater_than([1, 2, 3], value=0)
 
     """
     arr = arr if isinstance(arr, np.ndarray) else cast_to_ndarray(arr)
-    value = cast_to_ndarray(value)
-    check_has_shape(value, ())
-    check_is_real(value)
-    check_is_finite(value)
+    check_number(value)
+    check_real(value)
     if strict and not np.all(arr > value):
         raise ValueError(f"{name} values must all be greater than {value}.")
     elif not np.all(arr >= value):
         raise ValueError(f"{name} values must all be greater than or equal to {value}.")
 
 
-def check_is_less_than(arr, /, value, *, strict=True, name="Array"):
+def check_less_than(
+    arr: Union[_NumberType, Array[_NumberType]],
+    /,
+    value: float,
+    *,
+    strict: bool = True,
+    name: str = "Array",
+):
     """Check if an array's elements are all less than some value.
 
     Parameters
     ----------
-    arr : array_like
-        Array to check.
+    arr : float | Array[float]
+        Number or array to check.
 
-    value : Number
+    value : float
         Value which the array's elements must be less than.
 
     strict : bool, default: True
@@ -493,35 +516,45 @@ def check_is_less_than(arr, /, value, *, strict=True, name="Array"):
 
     See Also
     --------
-    check_is_greater_than
-    check_is_in_range
-    check_is_nonnegative
+    check_greater_than
+    check_range
+    check_nonnegative
 
     Examples
     --------
     Check if an array's values are less than 0.
 
-    >>> import pyvista.core.input_validation as valid
-    >>> valid.check_is_less_than([-1, -2, -3], value=0)
+    >>> from pyvista.core import validation
+    >>> validation.check_less_than([-1, -2, -3], value=0)
 
     """
     arr = arr if isinstance(arr, np.ndarray) else cast_to_ndarray(arr)
+    check_number(value)
+    check_real(value)
     if strict and not np.all(arr < value):
         raise ValueError(f"{name} values must all be less than {value}.")
     elif not np.all(arr <= value):
         raise ValueError(f"{name} values must all be less than or equal to {value}.")
 
 
-def check_is_in_range(arr, /, rng, *, strict_lower=False, strict_upper=False, name="Array"):
+def check_range(
+    arr: Union[_NumberType, Array[_NumberType]],
+    /,
+    rng: Vector[_NumberType],
+    *,
+    strict_lower: bool = False,
+    strict_upper: bool = False,
+    name: str = "Array",
+):
     """Check if an array's values are all within a specific range.
 
     Parameters
     ----------
-    arr : array_like
-        Array to check.
+    arr : float | Array[float]
+        Number or array to check.
 
-    rng : array_like[float, float], optional
-        Array-like with two elements ``[min, max]`` specifying the minimum
+    rng : Vector[float], optional
+        Vector with two elements ``[min, max]`` specifying the minimum
         and maximum data values allowed, respectively. By default, the
         range endpoints are inclusive, i.e. values must be >= min
         and <= max. Use ``strict_lower`` and/or ``strict_upper``
@@ -546,45 +579,44 @@ def check_is_in_range(arr, /, rng, *, strict_lower=False, strict_upper=False, na
 
     See Also
     --------
-    check_is_less_than
-    check_is_greater_than
+    check_less_than
+    check_greater_than
 
     Examples
     --------
     Check if `an array's values are in the range ``[0, 1]``.
 
-    >>> import pyvista.core.input_validation as valid
-    >>> valid.check_is_in_range([0, 0.5, 1], rng=[0, 1])
+    >>> from pyvista.core import validation
+    >>> validation.check_range([0, 0.5, 1], rng=[0, 1])
 
     """
-    rng = cast_to_ndarray(rng)
-    check_has_shape(rng, 2, name="Range")
-    check_is_real(rng, name="Range")
-    check_is_sorted(rng, name="Range")
+    rng_ = cast_to_ndarray(rng)
+    check_shape(rng_, 2, name="Range")
+    check_sorted(rng_, name="Range")
 
     arr = arr if isinstance(arr, np.ndarray) else cast_to_ndarray(arr)
     try:
-        check_is_greater_than(arr, rng[0], strict=strict_lower, name=name)
-        check_is_less_than(arr, rng[1], strict=strict_upper, name=name)
+        check_greater_than(arr, rng_[0], strict=strict_lower, name=name)
+        check_less_than(arr, rng_[1], strict=strict_upper, name=name)
     except ValueError:
         raise
 
 
-def check_has_shape(
-    arr,
+def check_shape(
+    arr: Union[_NumberType, Array[_NumberType]],
     /,
-    shape,
+    shape: Union[ShapeLike, List[ShapeLike]],
     *,
-    name="Array",
+    name: str = "Array",
 ):
     """Check if an array has the specified shape.
 
     Parameters
     ----------
-    arr : array_like
-        Array to check.
+    arr : float | Array[float]
+        Number or array to check.
 
-    shape : int, tuple[int, ...] | list[int, tuple[int, ...]], optional
+    shape : ShapeLike | list[ShapeLike], optional
         A single shape or a list of any allowable shapes. If an integer,
         ``i``, the shape is interpreted as ``(i,)``. Use a value of
         -1 for any dimension where its size is allowed to vary, e.g.
@@ -602,27 +634,27 @@ def check_has_shape(
 
     See Also
     --------
-    check_has_length
+    check_length
 
     Examples
     --------
-    Check if an array is one-dimensional.
+    Check if an array is one-dimension
 
     >>> import numpy as np
-    >>> import pyvista.core.input_validation as valid
-    >>> valid.check_has_shape([1, 2, 3], shape=(-1))
+    >>> from pyvista.core import validation
+    >>> validation.check_shape([1, 2, 3], shape=(-1))
 
     Check if an array is one-dimensional or a scalar.
 
-    >>> valid.check_has_shape(1, shape=[(), (-1)])
+    >>> validation.check_shape(1, shape=[(), (-1)])
 
     Check if an array is 3x3 or 4x4.
 
-    >>> valid.check_has_shape(np.eye(3), shape=[(3, 3), (4, 4)])
+    >>> validation.check_shape(np.eye(3), shape=[(3, 3), (4, 4)])
 
     """
 
-    def _shape_is_allowed(a, b):
+    def _shape_is_allowed(a: Shape, b: Shape) -> bool:
         # a: array's actual shape
         # b: allowed shape (may have -1)
         if len(a) == len(b) and all(map(lambda x, y: True if x == y else y == -1, a, b)):
@@ -636,9 +668,9 @@ def check_has_shape(
         shape = [shape]
 
     array_shape = arr.shape
-    for shp in shape:
-        shp = _validate_shape_value(shp)
-        if _shape_is_allowed(array_shape, shp):
+    for input_shape in shape:
+        input_shape = _validate_shape_value(input_shape)
+        if _shape_is_allowed(array_shape, input_shape):
             return
 
     msg = f"{name} has shape {arr.shape} which is not allowed. "
@@ -649,20 +681,61 @@ def check_has_shape(
     raise ValueError(msg)
 
 
-def check_is_number(num, /, *, name='Object'):
-    """Check if an object is an instance of ``Number``.
+def check_number(
+    num: Union[float, int, complex, np.number, Number],
+    /,
+    *,
+    definition: Literal['abstract', 'builtin', 'numpy'] = 'abstract',
+    must_be_real=True,
+    name: str = 'Object',
+):
+    """Check if an object is a number.
 
-    A number is any instance of ``numbers.Number``, e.g.  ``int``,
-    ``float``, and ``complex``.
+    By default, the number must be an instance of the abstract base class :class:`numbers.Real`.
+    Optionally, the number can also be complex. The definition can also be restricted
+    to strictly check if the number is a built-in numeric type (e.g. ``int``, ``float``)
+    or numpy numeric data types (e.g. ``np.floating``, ``np.integer``).
 
     Notes
     -----
-    A NumPy ndarray is not an instance of ``Number``.
+    - This check fails for instances of :class:`numpy.ndarray`. Use :func:`check_scalar`
+      instead to also allow for 0-dimensional arrays.
+    - Values such as ``float('inf')`` and ``float('NaN')`` are valid numbers and
+      will not raise an error. Use :func:`check_finite` to check for finite values.
+
+    .. warning::
+
+        - Some NumPy numeric data types are subclasses of the built-in types whereas other are
+          not. For example, ``numpy.float_`` is a subclass of ``float`` and ``numpy.complex_``
+          is a subclass ``complex``. However, ``numpy.int_`` is not a subclass of ``int`` and
+          ``numpy.bool_`` is not a subclass of ``bool``.
+        - The built-in ``bool`` type is a subclass of ``int`` whereas NumPy's``.bool_`` type
+          is not a subclass of ``np.int_`` (``np.bool_`` is not a numeric type).
+
+        This can lead to unexpected results:
+
+        - This check will always fail for ``np.bool_`` types.
+        - This check will pass for ``np.float_`` or ``np.complex_``, even if
+          ``definition='builtin'``.
 
     Parameters
     ----------
-    num : Number
+    num : float | int | complex | numpy.number | Number
         Number to check.
+
+    definition : str, default: 'abstract'
+        Control the base class(es) to use when checking the number's type. Must be
+        one of:
+
+        - ``'abstract'`` : number must be an instance of one of the abstract types
+          in :py:mod:`numbers`.
+        - ``'builtin'`` : number must be an instance of one of the built-in numeric
+          types.
+        - ``'numpy'`` : number must be an instance of NumPy's data types.
+
+    must_be_real : bool, default: True
+        If ``True``, the number must be real, i.e. an integer or
+        floating type. Set to ``False`` to allow complex numbers.
 
     name : str, default: "Object"
         Variable name to use in the error messages if any are raised.
@@ -670,27 +743,63 @@ def check_is_number(num, /, *, name='Object'):
     Raises
     ------
     TypeError
-        If input is not an instance of ``Number``.
+        If input is not an instance of a numeric type.
 
     See Also
     --------
-    check_is_scalar
+    check_scalar
+        Similar function which allows 0-dimensional ndarrays.
+    check_numeric
+        Similar function for any dimensional array of numbers.
+    check_real
+        Similar function for any dimensional array of real numbers.
+    check_finite
+
 
     Examples
     --------
-    Check if a complex number is an instance of ``Number``.
+    Check if a float is a number.
+    >>> from pyvista.core import validation
+    >>> num = 42.0
+    >>> type(num)
+    <class 'float'>
+    >>> validation.check_number(num)
 
-    >>> import pyvista.core.input_validation as valid
-    >>> valid.check_is_number(1 + 2j)
+    Check if an element of a NumPy array is a number.
+
+    >>> import numpy as np
+    >>> num_array = np.array([1, 2, 3])
+    >>> num = num_array[0]
+    >>> type(num)
+    <class 'numpy.int64'>
+    >>> validation.check_number(num)
+
+    Check if a complex number is a number.
+    >>> num = 1 + 2j
+    >>> type(num)
+    <class 'complex'>
+    >>> validation.check_number(num, must_be_real=False)
 
     """
+    check_contains(definition, ['abstract', 'builtin', 'numpy'])
+
+    valid_type: Any
+    if definition == 'abstract':
+        valid_type = Real if must_be_real else Number
+    elif definition == 'builtin':
+        valid_type = (float, int) if must_be_real else (float, int, complex)
+    elif definition == 'numpy':
+        valid_type = (np.floating, np.integer) if must_be_real else np.number
+    else:
+        raise NotImplementedError  # pragma: no cover
+
     try:
-        check_is_instance(num, Number, allow_subclass=True, name=name)
+        check_instance(num, valid_type, allow_subclass=True, name=name)
     except TypeError:
         raise
 
 
-def check_is_string(obj, /, *, allow_subclass=True, name='Object'):
+def check_string(obj: str, /, *, allow_subclass: bool = True, name: str = 'Object'):
     """Check if an object is an instance of ``str``.
 
     Parameters
@@ -712,26 +821,26 @@ def check_is_string(obj, /, *, allow_subclass=True, name='Object'):
 
     See Also
     --------
-    check_is_string_in_iterable
-    check_is_iterable_of_strings
-    check_is_sequence
-    check_is_instance
+    check_contains
+    check_iterable_items
+    check_sequence
+    check_instance
 
     Examples
     --------
     Check if an object is a string.
 
-    >>> import pyvista.core.input_validation as valid
-    >>> valid.check_is_string("eggs")
+    >>> from pyvista.core import validation
+    >>> validation.check_string("eggs")
 
     """
     try:
-        check_is_instance(obj, str, allow_subclass=allow_subclass, name=name)
+        check_instance(obj, str, allow_subclass=allow_subclass, name=name)
     except TypeError:
         raise
 
 
-def check_is_sequence(obj, /, *, name='Object'):
+def check_sequence(obj: Sequence, /, *, name: str = 'Object'):
     """Check if an object is an instance of ``Sequence``.
 
     Parameters
@@ -749,26 +858,26 @@ def check_is_sequence(obj, /, *, name='Object'):
 
     See Also
     --------
-    check_is_iterable
-    check_is_instance
+    check_iterable
+    check_instance
 
     Examples
     --------
     Check if an object is a sequence.
 
     >>> import numpy as np
-    >>> import pyvista.core.input_validation as valid
-    >>> valid.check_is_sequence([1, 2, 3])
-    >>> valid.check_is_sequence("A")
+    >>> from pyvista.core import validation
+    >>> validation.check_sequence([1, 2, 3])
+    >>> validation.check_sequence("A")
 
     """
     try:
-        check_is_instance(obj, Sequence, allow_subclass=True, name=name)
+        check_instance(obj, Sequence, allow_subclass=True, name=name)
     except TypeError:
         raise
 
 
-def check_is_iterable(obj, /, *, name='Object'):
+def check_iterable(obj: Iterable, /, *, name: str = 'Object'):
     """Check if an object is an instance of ``Iterable``.
 
     Parameters
@@ -786,27 +895,34 @@ def check_is_iterable(obj, /, *, name='Object'):
 
     See Also
     --------
-    check_is_sequence
-    check_is_instance
-    check_is_iterable_of_some_type
+    check_sequence
+    check_instance
+    check_iterable_items
 
     Examples
     --------
     Check if an object is iterable.
 
     >>> import numpy as np
-    >>> import pyvista.core.input_validation as valid
-    >>> valid.check_is_iterable([1, 2, 3])
-    >>> valid.check_is_iterable(np.array((4, 5, 6)))
+    >>> from pyvista.core import validation
+    >>> validation.check_iterable([1, 2, 3])
+    >>> validation.check_iterable(np.array((4, 5, 6)))
 
     """
     try:
-        check_is_instance(obj, Iterable, allow_subclass=True, name=name)
+        check_instance(obj, Iterable, allow_subclass=True, name=name)
     except TypeError:
         raise
 
 
-def check_is_instance(obj, /, classinfo, *, allow_subclass=True, name='Object'):
+def check_instance(
+    obj: Any,
+    /,
+    classinfo: Union[type, Tuple[type, ...]],
+    *,
+    allow_subclass: bool = True,
+    name: str = 'Object',
+):
     """Check if an object is an instance of the given type or types.
 
     Parameters
@@ -815,12 +931,12 @@ def check_is_instance(obj, /, classinfo, *, allow_subclass=True, name='Object'):
         Object to check.
 
     classinfo : type | tuple[type, ...]
-        ``type`` or tuple of types. Object must be an instance of one of
-        the types.
+        A type, tuple of types, or a Union of types. Object must be an instance
+        of one of the types.
 
     allow_subclass : bool, default: True
         If ``True``, the object's type must be specified by ``classinfo``
-         or any of its subclasses. Otherwise, subclasses are not allowed.
+        or any of its subclasses. Otherwise, subclasses are not allowed.
 
     name : str, default: "Object"
         Variable name to use in the error messages if any are raised.
@@ -832,22 +948,22 @@ def check_is_instance(obj, /, classinfo, *, allow_subclass=True, name='Object'):
 
     See Also
     --------
-    check_is_type
-    check_is_number
-    check_is_string
-    check_is_iterable
-    check_is_sequence
+    check_type
+    check_number
+    check_string
+    check_iterable
+    check_sequence
 
     Examples
     --------
     Check if an object is an instance of ``complex``.
 
-    >>> import pyvista.core.input_validation as valid
-    >>> valid.check_is_instance(1 + 2j, complex)
+    >>> from pyvista.core import validation
+    >>> validation.check_instance(1 + 2j, complex)
 
     Check if an object is an instance of one of several types.
 
-    >>> valid.check_is_instance("eggs", (int, str))
+    >>> validation.check_instance("eggs", (int, str))
 
     """
     if not isinstance(name, str):
@@ -858,7 +974,7 @@ def check_is_instance(obj, /, classinfo, *, allow_subclass=True, name='Object'):
         classinfo = get_args(classinfo)
 
     # Count num classes
-    if isinstance(classinfo, tuple):
+    if isinstance(classinfo, tuple) and all(isinstance(cls, type) for cls in classinfo):
         num_classes = len(classinfo)
     else:
         num_classes = 1
@@ -890,13 +1006,13 @@ def check_is_instance(obj, /, classinfo, *, allow_subclass=True, name='Object'):
         raise TypeError(msg)
 
 
-def check_is_type(obj, /, classinfo, *, name='Object'):
+def check_type(obj: Any, /, classinfo: Union[type, Tuple[type, ...]], *, name: str = 'Object'):
     """Check if an object is one of the given type or types.
 
     Notes
     -----
-    The use of :func:`check_is_instance` is generally preferred as it
-    allows subclasses. Use :func:`check_is_type` only for cases where
+    The use of :func:`check_instance` is generally preferred as it
+    allows subclasses. Use :func:`check_type` only for cases where
     exact types are necessary.
 
     Parameters
@@ -905,7 +1021,8 @@ def check_is_type(obj, /, classinfo, *, name='Object'):
         Object to check.
 
     classinfo : type | tuple[type, ...]
-        ``type`` or tuple of types. Object must be one of the types.
+        A type, tuple of types, or a Union of types. Object must be one
+        of the types.
 
     name : str, default: "Object"
         Variable name to use in the error messages if any are raised.
@@ -917,29 +1034,29 @@ def check_is_type(obj, /, classinfo, *, name='Object'):
 
     See Also
     --------
-    check_is_instance
+    check_instance
 
     Examples
     --------
     Check if an object is type ``dict`` or ``set``.
 
-    >>> import pyvista.core.input_validation as valid
-    >>> valid.check_is_type({'spam': "eggs"}, (dict, set))
+    >>> from pyvista.core import validation
+    >>> validation.check_type({'spam': "eggs"}, (dict, set))
 
     """
     try:
-        check_is_instance(obj, classinfo, allow_subclass=False, name=name)
+        check_instance(obj, classinfo, allow_subclass=False, name=name)
     except TypeError:
         raise
 
 
-def check_is_iterable_of_some_type(
-    iterable_obj,
+def check_iterable_items(
+    iterable_obj: Iterable,
     /,
-    some_type,
+    item_type: Union[type, Tuple[type, ...]],
     *,
-    allow_subclass=True,
-    name='Iterable',
+    allow_subclass: bool = True,
+    name: str = 'Iterable',
 ):
     """Check if an iterable's items all have a specified type.
 
@@ -948,12 +1065,12 @@ def check_is_iterable_of_some_type(
     iterable_obj : Iterable
         Iterable to check.
 
-    some_type : type | tuple[type, ...]
+    item_type : type | tuple[type, ...]
         Class type(s) to check for. Each element of the sequence must
         have the type or one of the types specified.
 
     allow_subclass : bool, default: True
-        If ``True``, the type of the iterable items must be any of the
+        If ``True``, the type of the iterable's items must be any of the
         given types or a subclass thereof. Otherwise, subclasses are not
         allowed.
 
@@ -967,88 +1084,48 @@ def check_is_iterable_of_some_type(
 
     See Also
     --------
-    check_is_instance
-    check_is_iterable
-    check_is_iterable_of_strings
+    check_instance
+    check_iterable
+    check_iterable_items
 
     Examples
     --------
     Check if a ``tuple`` only has ``int`` or ``float`` elements.
 
-    >>> import pyvista.core.input_validation as valid
-    >>> valid.check_is_iterable_of_some_type((1, 2, 3.0), (int, float))
+    >>> from pyvista.core import validation
+    >>> validation.check_iterable_items((1, 2, 3.0), (int, float))
 
     Check if a ``list`` only has ``list`` elements.
 
-    >>> import pyvista.core.input_validation as valid
-    >>> valid.check_is_iterable_of_some_type([[1], [2], [3]], list)
+    >>> from pyvista.core import validation
+    >>> validation.check_iterable_items([[1], [2], [3]], list)
 
     """
-    check_is_iterable(iterable_obj, name=name)
+    check_iterable(iterable_obj, name=name)
     try:
-        [
-            check_is_instance(
-                item, some_type, allow_subclass=allow_subclass, name=f"All items of {name}"
+        any(
+            check_instance(
+                item, item_type, allow_subclass=allow_subclass, name=f"All items of {name}"
             )
             for item in iterable_obj
-        ]
+        )
+
     except TypeError:
         raise
 
 
-def check_is_iterable_of_strings(iterable_obj, /, *, allow_subclass=True, name='String Iterable'):
-    """Check if an iterable's items are all strings.
+def check_contains(obj: Any, /, container: Any, *, name: str = 'Input'):
+    """Check if an object is in a container.
 
     Parameters
     ----------
-    iterable_obj : Iterable
-        Iterable to check.
+    obj : Any
+        Object to check.
 
-    allow_subclass : bool, default: True
-        If ``True``, the type of the iterable's items must be any of the
-        given types or a subclass thereof. Otherwise, subclasses are not
-        allowed.
+    container : Any
+        Container the object is expected to be in.
 
-    name : str, default: "String Iterable"
-        Variable name to use in the error messages if any are raised.
-
-    Raises
-    ------
-    TypeError
-        If any of the elements have an incorrect type.
-
-    See Also
-    --------
-    check_is_iterable
-    check_is_string
-    check_is_string_in_iterable
-
-    Examples
-    --------
-    Check if a ``tuple`` only has ``str`` elements.
-
-    >>> import pyvista.core.input_validation as valid
-    >>> valid.check_is_iterable_of_strings(("cat", "dog"))
-
-    """
-    try:
-        check_is_iterable_of_some_type(iterable_obj, str, allow_subclass=allow_subclass, name=name)
-    except TypeError:
-        raise
-
-
-def check_is_string_in_iterable(string_in, /, string_iterable, *, name='String'):
-    """Check if a given string is in an iterable of strings.
-
-    Parameters
-    ----------
-    string_in : str
-        String to check.
-
-    string_iterable : Iterable[str, ...]
-        Iterable containing only strings.
-
-    name : str, default: "String"
+    name : str, default: "Input"
         Variable name to use in the error messages if any are raised.
 
     Raises
@@ -1058,37 +1135,36 @@ def check_is_string_in_iterable(string_in, /, string_iterable, *, name='String')
 
     See Also
     --------
-    check_is_iterable
-    check_is_string
-    check_is_iterable_of_strings
+    check_iterable
+    check_iterable_items
 
     Examples
     --------
     Check if ``"A"`` is in a list of strings.
 
-    >>> import pyvista.core.input_validation as valid
-    >>> valid.check_is_string_in_iterable("A", ["A", "B", "C"])
+    >>> from pyvista.core import validation
+    >>> check_contains("A", ["A", "B", "C"])
 
     """
-    check_is_string(string_in, name=name)
-    check_is_iterable_of_strings(string_iterable)
-    if string_in not in string_iterable:
-        raise ValueError(
-            f"{name} '{string_in}' is not in the iterable. "
-            f"{name} must be one of: \n\t" + str(string_iterable)
-        )
+    if obj not in container:
+        if isinstance(container, (list, tuple)):
+            qualifier = "one of"
+        else:
+            qualifier = "in"
+        msg = f"{name} '{obj}' is not valid. {name} must be " f"{qualifier}: \n\t{container}"
+        raise ValueError(msg)
 
 
-def check_has_length(
-    arr,
+def check_length(
+    arr: Union[_NumberType, Array[_NumberType], Sized],
     /,
     *,
-    exact_length=None,
-    min_length=None,
-    max_length=None,
-    must_be_1d=False,
-    allow_scalars=False,
-    name="Array",
+    exact_length: Union[int, Vector[int], None] = None,
+    min_length: Optional[int] = None,
+    max_length: Optional[int] = None,
+    must_be_1d: bool = False,
+    allow_scalar: bool = False,
+    name: str = "Array",
 ):
     """Check if the length of an array meets specific requirements.
 
@@ -1101,10 +1177,10 @@ def check_has_length(
 
     Parameters
     ----------
-    arr : array_like
-        Array to check.
+    arr : float | Array[float]
+        Number or array to check.
 
-    exact_length : int | array_like[int, ...]
+    exact_length : int | Vector[int], optional
         Check if the array has the given length. If multiple
         numbers are given, the array's length must match one of the
         numbers.
@@ -1119,9 +1195,9 @@ def check_has_length(
         If ``True``, check if the shape of the array is one-dimensional,
         i.e. that the array's shape is ``(1,)``.
 
-    allow_scalars : bool, default: False
-        If ``True``, a scalar input will be reshaped to have a length of
-        1. Otherwise, the check will fail since a scalar does not
+    allow_scalar : bool, default: False
+        If ``True``, a scalar input will be reshaped to have a length
+        of 1. Otherwise, the check will fail since a scalar does not
         have a length.
 
     name : str, default: "Array"
@@ -1134,71 +1210,71 @@ def check_has_length(
 
     See Also
     --------
-    check_has_shape
+    check_shape
 
     Examples
     --------
     Check if an array has a length of 2 or 3.
 
-    >>> import pyvista.core.input_validation as valid
-    >>> valid.check_has_length([1, 2], exact_length=[2, 3])
+    >>> from pyvista.core import validation
+    >>> validation.check_length([1, 2], exact_length=[2, 3])
 
     Check if an array has a minimum length of 3.
 
-    >>> valid.check_has_length((1, 2, 3), min_length=3)
+    >>> validation.check_length((1, 2, 3), min_length=3)
 
     Check if a multidimensional array has a maximum length of 2.
 
-    >>> valid.check_has_length([[1, 2, 3], [4, 5, 6]], max_length=2)
+    >>> validation.check_length([[1, 2, 3], [4, 5, 6]], max_length=2)
 
     """
-    if allow_scalars:
+    if allow_scalar:
         # Reshape to 1D
-        if isinstance(arr, np.ndarray) and arr.ndim == 0:
-            arr = [arr.tolist()]
-        elif isinstance(arr, Number):
-            arr = [arr]
-    check_is_instance(arr, (Sequence, np.ndarray), name=name)
+        if isinstance(arr, Number):
+            arr = np.array([arr])
+        elif isinstance(arr, np.ndarray) and arr.ndim == 0:
+            arr = arr.reshape((1,))
+
+    check_instance(arr, (Sequence, np.ndarray), name=name)
+    arr = cast(Union[_NumberType, Array[_NumberType]], arr)
 
     if must_be_1d:
-        check_has_shape(arr, shape=(-1))
+        check_shape(arr, shape=(-1))
+
+    arr_len = len(cast(Sized, arr))
 
     if exact_length is not None:
-        exact_length = np.array(exact_length)
-        check_is_integerlike(exact_length, name="'exact_length'")
-        if len(arr) not in exact_length:
+        exact_length = cast_to_ndarray(exact_length)
+        check_integerlike(exact_length, name="'exact_length'")
+        if arr_len not in exact_length:
             raise ValueError(
                 f"{name} must have a length equal to any of: {exact_length}. "
-                f"Got length {len(arr)} instead."
+                f"Got length {arr_len} instead."
             )
 
     # Validate min/max length
     if min_length is not None:
-        min_length = cast_to_ndarray(min_length)
-        check_is_scalar(min_length, name="Min length")
-        check_is_real(min_length, name="Min length")
+        check_finite(min_length, name="Min length")
     if max_length is not None:
-        max_length = cast_to_ndarray(max_length)
-        check_is_scalar(max_length, name="Max length")
-        check_is_real(max_length, name="Max length")
+        check_finite(max_length, name="Max length")
     if min_length is not None and max_length is not None:
-        check_is_sorted((min_length, max_length), name="Range")
+        check_sorted((min_length, max_length), name="Range")
 
     if min_length is not None:
-        if len(arr) < min_length:
+        if arr_len < min_length:
             raise ValueError(
                 f"{name} must have a minimum length of {min_length}. "
-                f"Got length {len(arr)} instead."
+                f"Got length {arr_len} instead."
             )
     if max_length is not None:
-        if len(arr) > max_length:
+        if arr_len > max_length:
             raise ValueError(
                 f"{name} must have a maximum length of {max_length}. "
-                f"Got length {len(arr)} instead."
+                f"Got length {arr_len} instead."
             )
 
 
-def _validate_shape_value(shape: Union[int, Tuple[int, ...], Tuple[None]]):
+def _validate_shape_value(shape: ShapeLike) -> Shape:
     """Validate shape-like input and return its tuple representation."""
     if shape is None:
         # `None` is used to mean `any shape is allowed` by the array
@@ -1207,35 +1283,58 @@ def _validate_shape_value(shape: Union[int, Tuple[int, ...], Tuple[None]]):
         raise TypeError("`None` is not a valid shape. Use `()` instead.")
 
     # Return early for common inputs
-    if shape in [(), (-1,), (1,), (3,), (2,), (1, 3), (-1, 3)]:
-        return shape
+    if shape == ():
+        return cast(Shape, shape)
+    elif shape in [(-1,), (1,), (3,), (2,), (1, 3), (-1, 3)]:
+        return cast(Shape, shape)
 
     def _is_valid_dim(d):
         return isinstance(d, int) and d >= -1
 
     if _is_valid_dim(shape):
-        return (shape,)
+        return (cast(int, shape),)
     if isinstance(shape, tuple) and all(map(_is_valid_dim, shape)):
-        return shape
+        return cast(Shape, shape)
 
     # Input is not valid at this point. Use checks to raise an
     # appropriate error
-    check_is_instance(shape, (int, tuple), name='Shape')
+    check_instance(shape, (int, tuple), name='Shape')
     if isinstance(shape, int):
         shape = (shape,)
     else:
-        check_is_iterable_of_some_type(shape, int, name='Shape')
-    check_is_greater_than(shape, -1, name="Shape", strict=False)
+        check_iterable_items(shape, int, name='Shape')
+    check_greater_than(shape, -1, name="Shape", strict=False)
     raise RuntimeError("This line should not be reachable.")  # pragma: no cover
 
 
-def check_is_scalar(scalar, /, *, name="Scalar"):
-    """Check if an object is a real-valued scalar number.
+def check_scalar(
+    scalar: Union[float, int, complex, Number, NumpyArray[_NumberType]],
+    /,
+    *,
+    must_be_real: bool = True,
+    name: str = "Scalar",
+):
+    """Check if an object is a scalar number.
+
+    By default, the number must be real, or a 0-dimensional :class:`numpy.ndarray`
+    of a real number. Optionally, the scalar can also be a complex
+
+    This check is similar to :func:`check_number` but also allows 0-dimensional
+    arrays.
+
+    Notes
+    -----
+    Values such as ``infinity`` and ``NaN`` are valid scalars and will not
+    raise an error. Use :func:`check_finite` to check for finite values.
 
     Parameters
     ----------
-    scalar : int | float | array_like[int] | array_like[float]
-        Real number as an ``int``, ``float``, or 0-dimensional array.
+    scalar : float | int | complex | Number | numpy.number | numpy.ndarray
+        Number or 0-dimensional numeric array.
+
+    must_be_real : bool, default: True
+        If ``True``, the scalar must be a real number, i.e. an integer or
+        floating type. Set to ``False`` to allow complex numbers.
 
     name : str, default: "Scalar"
         Variable name to use in the error messages if any are raised.
@@ -1243,27 +1342,38 @@ def check_is_scalar(scalar, /, *, name="Scalar"):
     Raises
     ------
     TypeError
-        If input is not an instance of ``Number`` of a 0-dimensional array.
+        If input is not a number or a 0-dimensional number array.
 
     See Also
     --------
-    check_is_number
-    check_is_real
+    check_number
+        Similar function which does not allow 0-dimensional ndarrays.
+    check_numeric
+        Similar function for any dimensional array of numbers.
+    check_real
+        Similar function for any dimensional array of real numbers.
+    check_finite
 
     Examples
     --------
     Check if an object is scalar.
 
     >>> import numpy as np
-    >>> import pyvista.core.input_validation as valid
-    >>> valid.check_is_scalar(0.0)
-    >>> valid.check_is_scalar(np.array(1))
+    >>> from pyvista.core import validation
+    >>> validation.check_scalar(0.0)
+    >>> validation.check_scalar(np.array(1))
+    >>> validation.check_scalar(np.array(1 + 2j), must_be_real=False)
 
     """
-    check_is_instance(scalar, (int, float, np.ndarray), name=name)
-    if isinstance(scalar, np.ndarray):
-        check_is_real(scalar, name=name)
-        if scalar.ndim > 0:
-            raise ValueError(
-                f"{name} must be a 0-dimensional array, got `ndim={scalar.ndim}` instead."
-            )
+    try:
+        if isinstance(scalar, np.ndarray):
+            # TODO: check_ndim(scalar, 0)
+            if scalar.ndim > 0:
+                raise ValueError(
+                    f"{name} must be a 0-dimensional array, got `ndim={scalar.ndim}` instead."
+                )
+            check_real(scalar, name=name) if must_be_real else check_numeric(scalar, name=name)  # type: ignore
+        else:
+            check_number(scalar, must_be_real=must_be_real)
+    except TypeError:
+        raise
